fix: cleanup code and structure
